Explicit stdlib dependencies: hijack TOML -> IR conversion
By desugaring stdlib dependencies this early in the pipeline, we
are able to touch much less of Cargo. On the other hand, "stdlib"
is no longer a valid version for the purpose of `[[replace]]`-ing.

This isn't so bad, because one can still replace the dep the if one
knows the version number of the normally-resolved stdlib version.
More philosophically, replacements are done *after* solving version
requirements, at which point version requirements are gone and only
concrete resolved versions remain. "stdlib", in delegating the
choice of version elsewhere, is thus more of a version requirement
than version.

TODO: add some CLI flags / .config options to control the
desugaring --- pruning of stdlib deps is hard-enabled